pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apastats"
version = "0.1.0"
description = "APA 7th edition compliant statistical tables and analyses for organizational science research, following Journal of Applied Psychology (JAP) norms."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Aman Kumar"},
]
keywords = [
    "apa", "jap", "organizational-behavior", "psychology",
    "statistics", "moderation", "mediation", "correlations",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.10",
    "statsmodels>=0.14",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
docx = ["python-docx>=1.0"]
cfa = ["semopy>=2.3"]
reliability = ["factor_analyzer>=0.5"]
all = ["python-docx>=1.0", "semopy>=2.3", "factor_analyzer>=0.5"]
dev = ["pytest>=7.0", "pytest-cov", "python-docx>=1.0", "semopy>=2.3", "factor_analyzer>=0.5"]

[tool.setuptools.packages.find]
where = ["."]
include = ["apastats*"]
